Clean up Nudge module structure
Split the large rules schema implementation into focused matcher modules for content, syntax, paths, project state, and URLs while keeping the public schema facade intact.

Extract the check command into smaller scan and issue-building steps so the command entrypoint reads as the workflow instead of carrying all rule traversal logic inline.

Share JSON hook merging between Claude and Codex setup to remove duplicated object/array handling, and update local architecture docs for the new schema layout.

Verification: cargo fmt --all; cargo test -p nudge; cargo clippy -p nudge --all-targets --all-features -- -D warnings
