+++ packages/nudge/src/cmd.rs
@@ -3,6 +3,7 @@
 pub mod check;
 pub mod claude;
 pub mod codex;
+pub(crate) mod json_hooks;
 pub mod syntaxtree;
 pub mod test;
 pub mod validate;
