+++ packages/nudge/src/rules/schema.rs
@@ -1,26 +1,25 @@
 //! Schema types for user-defined rules.
 
-use std::{
-    io::Write,
-    path::Path,
-    process::{Command, Stdio},
-    sync::{LazyLock, Mutex, MutexGuard},
-};
-
-use derive_more::Display;
-use glob::Pattern;
 use monostate::MustBe;
-use regex::Regex;
-use serde::{Deserialize, Deserializer, Serialize, Serializer};
-use tree_sitter::{
-    Language as TsLanguage, Parser, Query, QueryCursor, QueryError, StreamingIterator,
-};
+use serde::{Deserialize, Serialize};
 
 use crate::{
     fmap_match,
     snippet::{Annotation, Match, Span},
-    template::{self, Captures},
+    template,
 };
+
+pub use content::{ContentMatcher, RegexMatcher};
+pub use path::GlobMatcher;
+pub use project_state::ProjectStateMatcher;
+pub use syntax::{Language, TreeSitterQuery};
+pub use url::UrlMatcher;
+
+mod content;
+mod path;
+mod project_state;
+mod syntax;
+mod url;
 
 /// A rule configuration file.
 #[derive(Debug, Clone, Deserialize)]
@@ -162,17 +161,17 @@ impl From<&Rule> for Rule {
 ///       file: "**/*.rs"
 ///       content:
 ///         - kind: Regex
-///           pattern: "(?m)^[ \\t]+use "
+///           pattern: "(?m)^[ \t]+use "
 ///     - hook: PreToolUse
 ///       tool: Edit
 ///       file: "**/*.rs"
 ///       new_content:
 ///         - kind: Regex
-///           pattern: "(?m)^[ \\t]+use "
+///           pattern: "(?m)^[ \t]+use "
 /// ```
 ///
 /// We have three options to handle this:
-/// - Change the shape of the config file (not ideal as it's less ergonomic)
+/// - Change the shape of the config file (less ergonomic)
 /// - Use a custom deserializer (more complex)
 /// - Use `monostate` in combination with fields that are unused in the struct,
 ///   so that instead of trying to parse as a tagged union serde just tries all
@@ -296,525 +295,6 @@ pub struct PreToolUseBashMatcher {
     pub project_state: Vec<ProjectStateMatcher>,
 }
 
-/// The method used to match hook content.
-///
-/// Uses custom deserialization because the `SyntaxTree` variant needs to
-/// compile tree-sitter queries at parse time, which requires the `language`
-/// field to be available when processing the `query` field.
-#[derive(Debug, Clone, Serialize)]
-#[serde(tag = "kind")]
-pub enum ContentMatcher {
-    /// Match on a regular expression.
-    Regex {
-        /// The regex pattern to match.
-        pattern: RegexMatcher,
-
-        /// Optional suggestion template for this matcher.
-        ///
-        /// When provided, the suggestion is interpolated with the match's
-        /// capture groups and added to the match context as `suggestion`.
-        /// This can then be referenced in the rule's message as `{{ $suggestion
-        /// }}`.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        suggestion: Option<String>,
-    },
-
-    /// Match on a tree-sitter syntax query.
-    ///
-    /// Tree-sitter queries match against the AST structure of code, enabling
-    /// precise pattern matching that regex cannot achieve (e.g., matching
-    /// `use` statements only inside function bodies, not in `mod test` blocks).
-    ///
-    /// Query syntax: <https://tree-sitter.github.io/tree-sitter/using-parsers/queries>
-    SyntaxTree {
-        /// The language grammar to use for parsing.
-        ///
-        /// Required because tree-sitter queries must be compiled against a
-        /// specific grammar, and we validate queries at rule load time rather
-        /// than deferring to match time.
-        language: Language,
-
-        /// The tree-sitter query pattern.
-        query: TreeSitterQuery,
-
-        /// Optional suggestion template, same as Regex variant.
-        ///
-        /// Captures from the query (e.g., `@fn_name`) can be referenced as
-        /// `{{ $fn_name }}` in the suggestion template.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        suggestion: Option<String>,
-    },
-
-    /// Match using an external program.
-    ///
-    /// Runs the specified command with the content piped to stdin. If the
-    /// command exits with a non-zero status, the rule matches.
-    ///
-    /// This enables integration with external linters and formatters like
-    /// `markdownlint`, `prettier --check`, etc.
-    External {
-        /// The command to run, as a list of arguments.
-        ///
-        /// The first element is the program, subsequent elements are arguments.
-        /// The content being checked is piped to the command's stdin.
-        ///
-        /// Example: `["npx", "markdownlint", "--stdin"]`
-        command: Vec<String>,
-    },
-}
-
-impl<'de> Deserialize<'de> for ContentMatcher {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        // Intermediate representation that can deserialize all variants.
-        // We use raw strings for fields that need post-processing.
-        #[derive(Deserialize)]
-        struct Raw {
-            kind: String,
-            // Regex fields
-            pattern: Option<String>,
-            // SyntaxTree fields
-            language: Option<Language>,
-            query: Option<String>,
-            // External fields
-            command: Option<Vec<String>>,
-            // Shared field
-            suggestion: Option<String>,
-        }
-
-        let raw = Raw::deserialize(deserializer)?;
-
-        match raw.kind.as_str() {
-            "Regex" => {
-                let pattern_str = raw
-                    .pattern
-                    .ok_or_else(|| serde::de::Error::missing_field("pattern"))?;
-                let pattern = Regex::new(&pattern_str)
-                    .map(RegexMatcher)
-                    .map_err(serde::de::Error::custom)?;
-                Ok(ContentMatcher::Regex {
-                    pattern,
-                    suggestion: raw.suggestion,
-                })
-            }
-            "SyntaxTree" => {
-                let language = raw
-                    .language
-                    .ok_or_else(|| serde::de::Error::missing_field("language"))?;
-                let query_str = raw
-                    .query
-                    .ok_or_else(|| serde::de::Error::missing_field("query"))?;
-                let query =
-                    TreeSitterQuery::new(language, query_str).map_err(serde::de::Error::custom)?;
-                Ok(ContentMatcher::SyntaxTree {
-                    language,
-                    query,
-                    suggestion: raw.suggestion,
-                })
-            }
-            "External" => {
-                let command = raw
-                    .command
-                    .ok_or_else(|| serde::de::Error::missing_field("command"))?;
-                if command.is_empty() {
-                    return Err(serde::de::Error::custom("command cannot be empty"));
-                }
-                Ok(ContentMatcher::External { command })
-            }
-            other => Err(serde::de::Error::unknown_variant(
-                other,
-                &["Regex", "SyntaxTree", "External"],
-            )),
-        }
-    }
-}
-
-impl ContentMatcher {
-    /// Test whether this pattern matches a given string.
-    pub fn is_match(&self, s: &str) -> bool {
-        match self {
-            ContentMatcher::Regex { pattern, .. } => pattern.is_match(s),
-            ContentMatcher::SyntaxTree {
-                language, query, ..
-            } => {
-                let Some(tree) = language.parse(s) else {
-                    return false;
-                };
-                let mut cursor = QueryCursor::new();
-                let mut matches = cursor.matches(query.as_ref(), tree.root_node(), s.as_bytes());
-                matches.next().is_some()
-            }
-            ContentMatcher::External { command } => run_external_command(command, s).is_some(),
-        }
-    }
-
-    /// Get the spans of all matches in a given string.
-    ///
-    /// Spans are returned in the order of the matches, and are non-overlapping.
-    ///
-    /// `External` matchers are arbitrary programs that don't return span
-    /// context, so any such "match" returns a span covering the entire content.
-    pub fn matches(&self, s: &str) -> Vec<Span> {
-        match self {
-            ContentMatcher::Regex { pattern, .. } => pattern.matches(s),
-            ContentMatcher::SyntaxTree {
-                language, query, ..
-            } => {
-                let Some(tree) = language.parse(s) else {
-                    return Vec::new();
-                };
-                let mut cursor = QueryCursor::new();
-                let mut ts_matches = cursor.matches(query.as_ref(), tree.root_node(), s.as_bytes());
-                let mut spans = Vec::new();
-                while let Some(m) = ts_matches.next() {
-                    spans.push(union_of_captures(m));
-                }
-                spans
-            }
-            ContentMatcher::External { command } => {
-                if run_external_command(command, s).is_some() {
-                    vec![Span::from(0..s.len())]
-                } else {
-                    Vec::new()
-                }
-            }
-        }
-    }
-
-    /// Get matches with capture groups for template interpolation.
-    ///
-    /// If this matcher has a suggestion template, it will be interpolated
-    /// with the match's captures and added to the context as `suggestion`.
-    ///
-    /// `External` matchers are arbitrary programs that don't return span
-    /// context, so any such "match" returns a span covering the entire content.
-    pub fn matches_with_context(&self, s: &str) -> Vec<Match> {
-        match self {
-            ContentMatcher::Regex {
-                pattern,
-                suggestion,
-            } => {
-                let mut matches = pattern.matches_with_context(s);
-
-                // If there's a suggestion template, interpolate it per-match
-                if let Some(suggestion_template) = suggestion {
-                    for m in &mut matches {
-                        let interpolated = template::interpolate(suggestion_template, &m.captures);
-                        m.captures.insert("suggestion".to_string(), interpolated);
-                    }
-                }
-
-                matches
-            }
-            ContentMatcher::SyntaxTree {
-                language,
-                query,
-                suggestion,
-            } => {
-                let Some(tree) = language.parse(s) else {
-                    return Vec::new();
-                };
-
-                let mut cursor = QueryCursor::new();
-                let capture_names = query.as_ref().capture_names();
-                let mut ts_matches = cursor.matches(query.as_ref(), tree.root_node(), s.as_bytes());
-
-                let mut matches = Vec::new();
-                while let Some(m) = ts_matches.next() {
-                    let span = union_of_captures(m);
-                    let mut captures = Captures::new();
-
-                    // Extract named captures as source text, matching regex behavior.
-                    // This allows suggestions to reference captures like {{ $fn_name }}.
-                    for capture in m.captures {
-                        if let Some(name) = capture_names.get(capture.index as usize) {
-                            let text = capture
-                                .node
-                                .utf8_text(s.as_bytes())
-                                .unwrap_or_default()
-                                .to_string();
-                            captures.insert(name.to_string(), text);
-                        }
-                    }
-
-                    matches.push(Match { span, captures });
-                }
-
-                if let Some(suggestion_template) = suggestion {
-                    for m in &mut matches {
-                        let interpolated = template::interpolate(suggestion_template, &m.captures);
-                        m.captures.insert("suggestion".to_string(), interpolated);
-                    }
-                }
-
-                matches
-            }
-            ContentMatcher::External { command } => {
-                if let Some(command) = run_external_command(command, s) {
-                    let captures = Captures::from_iter([("command".to_string(), command)]);
-                    vec![Match {
-                        span: Span::from(0..s.len()),
-                        captures,
-                    }]
-                } else {
-                    Vec::new()
-                }
-            }
-        }
-    }
-}
-
-/// The method used to match URLs.
-///
-/// Similar to [`ContentMatcher`] but only supports regex patterns, since URLs
-/// are simple strings without syntax tree structure.
-#[derive(Debug, Clone, Serialize)]
-#[serde(tag = "kind")]
-pub enum UrlMatcher {
-    /// Match on a regular expression.
-    Regex {
-        /// The regex pattern to match against the URL.
-        pattern: RegexMatcher,
-
-        /// Optional suggestion template for this matcher.
-        ///
-        /// When provided, the suggestion is interpolated with the match's
-        /// capture groups and added to the match context as `suggestion`.
-        /// This can then be referenced in the rule's message as `{{ $suggestion
-        /// }}`.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        suggestion: Option<String>,
-    },
-}
-
-impl<'de> Deserialize<'de> for UrlMatcher {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        #[derive(Deserialize)]
-        struct Raw {
-            kind: String,
-            pattern: Option<String>,
-            suggestion: Option<String>,
-        }
-
-        let raw = Raw::deserialize(deserializer)?;
-
-        match raw.kind.as_str() {
-            "Regex" => {
-                let pattern_str = raw
-                    .pattern
-                    .ok_or_else(|| serde::de::Error::missing_field("pattern"))?;
-                let pattern = Regex::new(&pattern_str)
-                    .map(RegexMatcher)
-                    .map_err(serde::de::Error::custom)?;
-                Ok(UrlMatcher::Regex {
-                    pattern,
-                    suggestion: raw.suggestion,
-                })
-            }
-            other => Err(serde::de::Error::unknown_variant(other, &["Regex"])),
-        }
-    }
-}
-
-impl UrlMatcher {
-    /// Test whether this pattern matches a given URL.
-    pub fn is_match(&self, url: &str) -> bool {
-        match self {
-            UrlMatcher::Regex { pattern, .. } => pattern.is_match(url),
-        }
-    }
-
-    /// Get the spans of all matches in a given URL.
-    pub fn matches(&self, url: &str) -> Vec<Span> {
-        match self {
-            UrlMatcher::Regex { pattern, .. } => pattern.matches(url),
-        }
-    }
-
-    /// Get matches with capture groups for template interpolation.
-    ///
-    /// If this matcher has a suggestion template, it will be interpolated
-    /// with the match's captures and added to the context as `suggestion`.
-    pub fn matches_with_context(&self, url: &str) -> Vec<Match> {
-        match self {
-            UrlMatcher::Regex {
-                pattern,
-                suggestion,
-            } => {
-                let mut matches = pattern.matches_with_context(url);
-
-                // If there's a suggestion template, interpolate it per-match
-                if let Some(suggestion_template) = suggestion {
-                    for m in &mut matches {
-                        let interpolated = template::interpolate(suggestion_template, &m.captures);
-                        m.captures.insert("suggestion".to_string(), interpolated);
-                    }
-                }
-
-                matches
-            }
-        }
-    }
-}
-
-/// Matcher for project state conditions.
-///
-/// Project state matchers evaluate conditions about the project environment
-/// (e.g., git state) rather than the content of the tool input. All project
-/// state matchers in a rule must match for the rule to proceed.
-#[derive(Debug, Clone, Serialize)]
-#[serde(tag = "kind")]
-pub enum ProjectStateMatcher {
-    /// Match against git repository state.
-    Git {
-        /// Match against the current branch name.
-        ///
-        /// All branch matchers must match the current branch name for this
-        /// git matcher to pass.
-        #[serde(default)]
-        branch: Vec<ContentMatcher>,
-    },
-}
-
-impl<'de> Deserialize<'de> for ProjectStateMatcher {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        #[derive(Deserialize)]
-        struct Raw {
-            kind: String,
-            #[serde(default)]
-            branch: Vec<ContentMatcher>,
-        }
-
-        let raw = Raw::deserialize(deserializer)?;
-
-        match raw.kind.as_str() {
-            "Git" => Ok(ProjectStateMatcher::Git { branch: raw.branch }),
-            other => Err(serde::de::Error::unknown_variant(other, &["Git"])),
-        }
-    }
-}
-
-impl ProjectStateMatcher {
-    /// Evaluate this matcher against the project state at the given path.
-    ///
-    /// Returns `true` if the matcher matches, `false` otherwise.
-    /// Logs a warning if the project state cannot be determined (e.g., not
-    /// in a git repository).
-    pub fn is_match(&self, cwd: &Path) -> bool {
-        match self {
-            ProjectStateMatcher::Git { branch } => {
-                if branch.is_empty() {
-                    // No branch matchers = always match (just checking we're in a git repo)
-                    if crate::git::current_branch(cwd).is_some() {
-                        return true;
-                    }
-                    tracing::warn!(?cwd, "project_state.Git matcher: not in a git repository");
-                    return false;
-                }
-
-                let Some(current_branch) = crate::git::current_branch(cwd) else {
-                    tracing::warn!(
-                        ?cwd,
-                        "project_state.Git matcher: could not determine current branch"
-                    );
-                    return false;
-                };
-
-                // All branch matchers must match
-                branch.iter().all(|m| m.is_match(&current_branch))
-            }
-        }
-    }
-}
-
-/// Run an external command with content piped to stdin.
-///
-/// Returns `Some(formatted_command)` if the command exits with non-zero status
-/// (indicating a match/violation), or `None` if the command succeeds (no
-/// violation).
-fn run_external_command(command: &[String], content: &str) -> Option<String> {
-    let Some((program, args)) = command.split_first() else {
-        tracing::warn!("external command is empty");
-        return None;
-    };
-
-    let mut child = match Command::new(program)
-        .args(args)
-        .stdin(Stdio::piped())
-        .stdout(Stdio::null())
-        .stderr(Stdio::null())
-        .spawn()
-    {
-        Ok(child) => child,
-        Err(e) => {
-            tracing::warn!(?program, error = %e, "failed to spawn external command");
-            return None;
-        }
-    };
-
-    // Write content to stdin. We ignore write errors (e.g., EPIPE) because the
-    // command may exit before consuming all input (like `false` which exits
-    // immediately). What matters is the command's exit status, not whether it
-    // read all its input.
-    if let Some(mut stdin) = child.stdin.take() {
-        let _ = stdin.write_all(content.as_bytes());
-        // stdin is dropped here, closing the pipe
-    }
-
-    // Wait for the command to complete
-    match child.wait() {
-        Ok(status) => {
-            if status.success() {
-                // Command succeeded = no violation
-                None
-            } else {
-                // Command failed = violation detected
-                // Format the command for the hint message
-                Some(shell_words::join(command))
-            }
-        }
-        Err(e) => {
-            tracing::warn!(?program, error = %e, "failed to wait for external command");
-            None
-        }
-    }
-}
-
-/// Compute the union span of all captures in a tree-sitter match.
-///
-/// Tree-sitter matches can have multiple captures; we highlight the entire
-/// matched region rather than individual captures for consistency with how
-/// regex matches display the full match.
-fn union_of_captures(m: &tree_sitter::QueryMatch) -> Span {
-    if m.captures.is_empty() {
-        // Empty captures shouldn't happen with valid queries that have @captures,
-        // but return a placeholder span pointing to file start if it does.
-        tracing::warn!("tree-sitter match has no captures");
-        return Span { start: 0, end: 0 };
-    }
-
-    let (start, end) = m.captures.iter().fold((usize::MAX, 0), |(start, end), c| {
-        let range = c.node.byte_range();
-        (start.min(range.start), end.max(range.end))
-    });
-
-    // Defensive: ensure valid span even if captures have unexpected ranges
-    if start > end {
-        tracing::warn!("tree-sitter match produced invalid byte ranges");
-        return Span { start: 0, end: 0 };
-    }
-
-    Span { start, end }
-}
-
 /// Matches the UserPromptSubmit hook.
 #[derive(Debug, Clone, Deserialize, Serialize)]
 pub struct UserPromptSubmitMatcher {
@@ -829,609 +309,11 @@ pub struct UserPromptSubmitMatcher {
     pub prompt: Vec<ContentMatcher>,
 }
 
-/// Match on a glob pattern.
-#[derive(Debug, Clone, Display)]
-#[display("{_0}")]
-pub struct GlobMatcher(Pattern);
-
-impl GlobMatcher {
-    /// Create an instance that matches any string.
-    pub fn any() -> Self {
-        static ANY: LazyLock<Pattern> =
-            LazyLock::new(|| Pattern::new("**/*").expect("compile 'any' glob pattern"));
-        GlobMatcher(ANY.clone())
-    }
-
-    /// Test whether this pattern matches a given string.
-    pub fn is_match(&self, path: &str) -> bool {
-        self.0.matches(path)
-    }
-
-    /// Test whether this pattern matches the given path.
-    pub fn is_match_path(&self, path: &Path) -> bool {
-        self.0.matches_path(path)
-    }
-}
-
-impl Default for GlobMatcher {
-    fn default() -> Self {
-        Self::any()
-    }
-}
-
-impl<'de> Deserialize<'de> for GlobMatcher {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let s = String::deserialize(deserializer)?;
-        let pattern = Pattern::new(&s).map_err(serde::de::Error::custom)?;
-        Ok(GlobMatcher(pattern))
-    }
-}
-
-impl Serialize for GlobMatcher {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        serializer.serialize_str(&self.0.to_string())
-    }
-}
-
-/// Match on a regex pattern.
-#[derive(Debug, Clone, Display)]
-#[display("{_0}")]
-pub struct RegexMatcher(Regex);
-
-impl RegexMatcher {
-    /// Create an instance that matches any string.
-    pub fn any() -> Self {
-        static ANY: LazyLock<Regex> =
-            LazyLock::new(|| Regex::new(".*").expect("compile 'any' regex"));
-        RegexMatcher(ANY.clone())
-    }
-
-    /// Create an instance that matches nothing.
-    pub fn none() -> Self {
-        static NONE: LazyLock<Regex> =
-            LazyLock::new(|| Regex::new("a^").expect("compile 'none' regex"));
-        RegexMatcher(NONE.clone())
-    }
-
-    /// Test whether this pattern matches a given string.
-    pub fn is_match(&self, s: &str) -> bool {
-        self.0.is_match(s)
-    }
-
-    /// Get the spans of all matches in a given string.
-    ///
-    /// Spans are returned in the order of the matches, and are non-overlapping.
-    pub fn matches(&self, s: &str) -> Vec<Span> {
-        self.0.find_iter(s).map(|m| m.range().into()).collect()
-    }
-
-    /// Get matches with capture groups for template interpolation.
-    ///
-    /// Returns `Match` objects containing both the span and captured values.
-    /// Captures are stored as:
-    /// - `"0"`, `"1"`, `"2"`, etc. for positional captures
-    /// - Named keys for named capture groups (e.g., `"var_name"`)
-    pub fn matches_with_context(&self, s: &str) -> Vec<Match> {
-        self.0
-            .captures_iter(s)
-            .map(|caps| {
-                let full_match = caps.get(0).expect("capture 0 always exists");
-                let span = Span::from(full_match.range());
-
-                let mut captures = Captures::new();
-
-                // Add positional captures
-                for i in 0..caps.len() {
-                    if let Some(cap) = caps.get(i) {
-                        captures.insert(i.to_string(), cap.as_str().to_string());
-                    }
-                }
-
-                // Add named captures
-                for name in self.0.capture_names().flatten() {
-                    if let Some(cap) = caps.name(name) {
-                        captures.insert(name.to_string(), cap.as_str().to_string());
-                    }
-                }
-
-                Match { span, captures }
-            })
-            .collect()
-    }
-}
-
-impl Default for RegexMatcher {
-    fn default() -> Self {
-        Self::any()
-    }
-}
-
-impl<'de> Deserialize<'de> for RegexMatcher {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let s = String::deserialize(deserializer)?;
-        let regex = Regex::new(&s).map_err(serde::de::Error::custom)?;
-        Ok(RegexMatcher(regex))
-    }
-}
-
-impl Serialize for RegexMatcher {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        serializer.serialize_str(&self.0.to_string())
-    }
-}
-
-/// Supported languages for tree-sitter parsing.
-///
-/// Adding a new language requires:
-/// 1. Adding the grammar crate to Cargo.toml (e.g., `tree-sitter-python`)
-/// 2. Adding a variant to this enum
-/// 3. Adding a match arm to `grammar()` that returns the language
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, clap::ValueEnum)]
-#[serde(rename_all = "lowercase")]
-pub enum Language {
-    /// The Rust programming language.
-    Rust,
-    /// The TypeScript programming language.
-    TypeScript,
-    /// The JavaScript programming language.
-    JavaScript,
-    /// The Python programming language.
-    Python,
-    /// The Go programming language.
-    Go,
-    /// The Java programming language.
-    Java,
-    /// The C# programming language.
-    CSharp,
-    /// The Kotlin programming language.
-    Kotlin,
-    /// The Haskell programming language.
-    Haskell,
-}
-
-impl Language {
-    /// Get the tree-sitter grammar for this language.
-    pub fn grammar(self) -> TsLanguage {
-        match self {
-            Language::Rust => tree_sitter_rust::LANGUAGE.into(),
-            Language::TypeScript => tree_sitter_typescript::LANGUAGE_TYPESCRIPT.into(),
-            Language::JavaScript => tree_sitter_javascript::LANGUAGE.into(),
-            Language::Python => tree_sitter_python::LANGUAGE.into(),
-            Language::Go => tree_sitter_go::LANGUAGE.into(),
-            Language::Java => tree_sitter_java::LANGUAGE.into(),
-            Language::CSharp => tree_sitter_c_sharp::LANGUAGE.into(),
-            Language::Kotlin => tree_sitter_kotlin_ng::LANGUAGE.into(),
-            Language::Haskell => tree_sitter_haskell::LANGUAGE.into(),
-        }
-    }
-
-    /// Parse source code into a syntax tree.
-    ///
-    /// Returns `None` if parsing fails (e.g., malformed code). We intentionally
-    /// don't block on parse errors since code being written is often
-    /// incomplete.
-    pub fn parse(self, source: &str) -> Option<tree_sitter::Tree> {
-        // Reuse parsers across calls. Parser creation has non-trivial overhead,
-        // and parsers are designed to be reused. We use Mutex because parsing
-        // is stateful (the parser tracks incremental parse state).
-        static RUST_PARSER: LazyLock<Mutex<Parser>> = LazyLock::new(|| {
-            let mut parser = Parser::new();
-            parser
-                .set_language(&tree_sitter_rust::LANGUAGE.into())
-                .expect("failed to set Rust language");
-            Mutex::new(parser)
-        });
-
-        static TYPESCRIPT_PARSER: LazyLock<Mutex<Parser>> = LazyLock::new(|| {
-            let mut parser = Parser::new();
-            parser
-                .set_language(&tree_sitter_typescript::LANGUAGE_TYPESCRIPT.into())
-                .expect("failed to set TypeScript language");
-            Mutex::new(parser)
-        });
-
-        static JAVASCRIPT_PARSER: LazyLock<Mutex<Parser>> = LazyLock::new(|| {
-            let mut parser = Parser::new();
-            parser
-                .set_language(&tree_sitter_javascript::LANGUAGE.into())
-                .expect("failed to set JavaScript language");
-            Mutex::new(parser)
-        });
-
-        static PYTHON_PARSER: LazyLock<Mutex<Parser>> = LazyLock::new(|| {
-            let mut parser = Parser::new();
-            parser
-                .set_language(&tree_sitter_python::LANGUAGE.into())
-                .expect("failed to set Python language");
-            Mutex::new(parser)
-        });
-
-        static GO_PARSER: LazyLock<Mutex<Parser>> = LazyLock::new(|| {
-            let mut parser = Parser::new();
-            parser
-                .set_language(&tree_sitter_go::LANGUAGE.into())
-                .expect("failed to set Go language");
-            Mutex::new(parser)
-        });
-
-        static JAVA_PARSER: LazyLock<Mutex<Parser>> = LazyLock::new(|| {
-            let mut parser = Parser::new();
-            parser
-                .set_language(&tree_sitter_java::LANGUAGE.into())
-                .expect("failed to set Java language");
-            Mutex::new(parser)
-        });
-
-        static CSHARP_PARSER: LazyLock<Mutex<Parser>> = LazyLock::new(|| {
-            let mut parser = Parser::new();
-            parser
-                .set_language(&tree_sitter_c_sharp::LANGUAGE.into())
-                .expect("failed to set C# language");
-            Mutex::new(parser)
-        });
-
-        static KOTLIN_PARSER: LazyLock<Mutex<Parser>> = LazyLock::new(|| {
-            let mut parser = Parser::new();
-            parser
-                .set_language(&tree_sitter_kotlin_ng::LANGUAGE.into())
-                .expect("failed to set Kotlin language");
-            Mutex::new(parser)
-        });
-
-        static HASKELL_PARSER: LazyLock<Mutex<Parser>> = LazyLock::new(|| {
-            let mut parser = Parser::new();
-            parser
-                .set_language(&tree_sitter_haskell::LANGUAGE.into())
-                .expect("failed to set Haskell language");
-            Mutex::new(parser)
-        });
-
-        let mut parser = match self {
-            Language::Rust => lock_parser(&RUST_PARSER),
-            Language::TypeScript => lock_parser(&TYPESCRIPT_PARSER),
-            Language::JavaScript => lock_parser(&JAVASCRIPT_PARSER),
-            Language::Python => lock_parser(&PYTHON_PARSER),
-            Language::Go => lock_parser(&GO_PARSER),
-            Language::Java => lock_parser(&JAVA_PARSER),
-            Language::CSharp => lock_parser(&CSHARP_PARSER),
-            Language::Kotlin => lock_parser(&KOTLIN_PARSER),
-            Language::Haskell => lock_parser(&HASKELL_PARSER),
-        };
-
-        let tree = parser.parse(source, None)?;
-
-        // Log a warning if the tree has errors, but still return it.
-        // Partial parses can still match valid subtrees.
-        if tree.root_node().has_error() {
-            tracing::debug!(language = ?self, "parsed code contains syntax errors");
-        }
-
-        Some(tree)
-    }
-}
-
-fn lock_parser<T>(parser: &Mutex<T>) -> MutexGuard<'_, T> {
-    parser.lock().expect("parser mutex poisoned")
-}
-
-/// A compiled tree-sitter query.
-///
-/// Wraps `tree_sitter::Query` with the original source and language for
-/// serialization and cloning. Queries are compiled at deserialization time
-/// to catch errors early (at rule load time, not match time).
-#[derive(Debug)]
-pub struct TreeSitterQuery {
-    inner: Query,
-    source: String,
-    language: Language,
-}
-
-impl TreeSitterQuery {
-    /// Compile a query from source for the given language.
-    pub fn new(language: Language, source: impl Into<String>) -> Result<Self, QueryError> {
-        let source = source.into();
-        let inner = Query::new(&language.grammar(), &source)?;
-        Ok(Self {
-            inner,
-            source,
-            language,
-        })
-    }
-}
-
-impl AsRef<Query> for TreeSitterQuery {
-    fn as_ref(&self) -> &Query {
-        &self.inner
-    }
-}
-
-impl Clone for TreeSitterQuery {
-    fn clone(&self) -> Self {
-        // Safe to unwrap: if it compiled once, it will compile again
-        Self::new(self.language, &self.source).expect("query compiled before, should compile again")
-    }
-}
-
-impl Serialize for TreeSitterQuery {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        serializer.serialize_str(&self.source)
-    }
-}
-
-/// Deserialize a tree-sitter query.
-///
-/// This requires special handling because tree-sitter queries must be compiled
-/// against a language grammar. We use `deserialize_with` at the struct level
-/// to access the sibling `language` field.
-impl<'de> Deserialize<'de> for TreeSitterQuery {
-    fn deserialize<D>(_deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        // This impl exists only to satisfy the trait bound. Actual deserialization
-        // happens via the custom deserializer for ContentMatcher::SyntaxTree.
-        Err(serde::de::Error::custom(
-            "TreeSitterQuery cannot be deserialized standalone; use ContentMatcher::SyntaxTree",
-        ))
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use std::panic::catch_unwind;
-
     use pretty_assertions::assert_eq as pretty_assert_eq;
 
     use super::*;
-
-    #[test]
-    fn test_language_parse_valid_rust() {
-        let code = "fn main() { println!(\"hello\"); }";
-        let tree = Language::Rust.parse(code);
-        assert!(tree.is_some());
-    }
-
-    #[test]
-    fn test_language_parse_invalid_returns_tree_with_errors() {
-        // Tree-sitter is error-tolerant; it returns a tree even for invalid code
-        let code = "fn main( { }";
-        let tree = Language::Rust.parse(code);
-        assert!(tree.is_some());
-    }
-
-    #[test]
-    fn test_lock_parser_panics_when_mutex_is_poisoned() {
-        let parser = Mutex::new(());
-        let poison = catch_unwind(|| {
-            let _guard = parser.lock().expect("initial parser lock should succeed");
-            panic!("poison parser mutex");
-        });
-
-        assert!(poison.is_err());
-        assert!(parser.is_poisoned());
-
-        let result = catch_unwind(|| {
-            let _guard = lock_parser(&parser);
-        });
-
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_treesitter_query_compile_valid() {
-        let query = TreeSitterQuery::new(Language::Rust, "(function_item)");
-        assert!(query.is_ok());
-    }
-
-    #[test]
-    fn test_treesitter_query_compile_invalid() {
-        let query = TreeSitterQuery::new(Language::Rust, "(not_a_real_node)");
-        assert!(query.is_err());
-    }
-
-    #[test]
-    fn test_content_matcher_syntax_tree_deserialize() {
-        let yaml = r#"
-            kind: SyntaxTree
-            language: rust
-            query: "(function_item)"
-        "#;
-        let matcher = serde_yaml::from_str::<ContentMatcher>(yaml).expect("valid yaml");
-        assert!(matches!(matcher, ContentMatcher::SyntaxTree { .. }));
-    }
-
-    #[test]
-    fn test_content_matcher_syntax_tree_deserialize_invalid_query() {
-        let yaml = r#"
-            kind: SyntaxTree
-            language: rust
-            query: "(not_a_real_node)"
-        "#;
-        let result = serde_yaml::from_str::<ContentMatcher>(yaml);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_syntax_tree_is_match() {
-        let matcher = ContentMatcher::SyntaxTree {
-            language: Language::Rust,
-            query: TreeSitterQuery::new(Language::Rust, "(function_item)").expect("valid query"),
-            suggestion: None,
-        };
-        assert!(matcher.is_match("fn foo() {}"));
-        assert!(!matcher.is_match("let x = 1;"));
-    }
-
-    #[test]
-    fn test_syntax_tree_matches_returns_spans() {
-        let matcher = ContentMatcher::SyntaxTree {
-            language: Language::Rust,
-            query: TreeSitterQuery::new(Language::Rust, "(function_item) @fn")
-                .expect("valid tree-sitter query"),
-            suggestion: None,
-        };
-        let code = "fn foo() {}\nfn bar() {}";
-        let spans = matcher.matches(code);
-        pretty_assert_eq!(spans.len(), 2);
-    }
-
-    #[test]
-    fn test_syntax_tree_captures_as_source_text() {
-        let matcher = ContentMatcher::SyntaxTree {
-            language: Language::Rust,
-            query: TreeSitterQuery::new(
-                Language::Rust,
-                "(function_item body: (block (use_declaration) @use))",
-            )
-            .expect("valid tree-sitter query"),
-            suggestion: None,
-        };
-
-        // Should match: use inside function
-        let code_match = "fn foo() { use std::io; }";
-        assert!(matcher.is_match(code_match));
-
-        // Should NOT match: top-level use
-        let code_no_match = "use std::io;\nfn foo() {}";
-        assert!(!matcher.is_match(code_no_match));
-    }
-
-    #[test]
-    fn test_syntax_tree_malformed_code_passes() {
-        // Malformed code should not match (passes silently)
-        let matcher = ContentMatcher::SyntaxTree {
-            language: Language::Rust,
-            query: TreeSitterQuery::new(Language::Rust, "(function_item)")
-                .expect("valid tree-sitter query"),
-            suggestion: None,
-        };
-        // Completely malformed - no function keyword
-        let malformed = "{{{{";
-        // Tree-sitter is error-tolerant, but this won't match function_item
-        assert!(!matcher.is_match(malformed));
-    }
-
-    #[test]
-    fn test_union_of_captures_multiple() {
-        // Test that the union span covers all captures
-        let matcher = ContentMatcher::SyntaxTree {
-            language: Language::Rust,
-            query: TreeSitterQuery::new(
-                Language::Rust,
-                "(function_item name: (identifier) @name body: (block) @body)",
-            )
-            .expect("valid tree-sitter query"),
-            suggestion: None,
-        };
-        let code = "fn foo() { let x = 1; }";
-        let spans = matcher.matches(code);
-        pretty_assert_eq!(spans.len(), 1);
-        // The span should cover from "foo" through the end of the block
-        let matched_text = &code[spans[0].start..spans[0].end];
-        assert!(matched_text.contains("foo"));
-        assert!(matched_text.contains("let x = 1"));
-    }
-
-    #[test]
-    fn test_external_deserialize() {
-        let yaml = r#"
-            kind: External
-            command: ["grep", "-q", "error"]
-        "#;
-        let matcher =
-            serde_yaml::from_str::<ContentMatcher>(yaml).expect("valid external matcher yaml");
-        assert!(matches!(matcher, ContentMatcher::External { .. }));
-    }
-
-    #[test]
-    fn test_external_deserialize_empty_command() {
-        let yaml = r#"
-            kind: External
-            command: []
-        "#;
-        let result = serde_yaml::from_str::<ContentMatcher>(yaml);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_external_is_match_when_command_fails() {
-        // `false` is a command that always exits with code 1
-        let matcher = ContentMatcher::External {
-            command: vec!["false".to_string()],
-        };
-        assert!(matcher.is_match("any content"));
-    }
-
-    #[test]
-    fn test_external_is_not_match_when_command_succeeds() {
-        // `true` is a command that always exits with code 0
-        let matcher = ContentMatcher::External {
-            command: vec!["true".to_string()],
-        };
-        assert!(!matcher.is_match("any content"));
-    }
-
-    #[test]
-    fn test_external_matches_with_context_sets_command_capture() {
-        // `false` always fails
-        let matcher = ContentMatcher::External {
-            command: vec!["false".to_string()],
-        };
-        let matches = matcher.matches_with_context("content");
-        pretty_assert_eq!(matches.len(), 1);
-        pretty_assert_eq!(
-            matches[0].captures.get("command"),
-            Some(&"false".to_string())
-        );
-    }
-
-    #[test]
-    fn test_external_matches_with_context_formats_command_with_args() {
-        // `test 1 -eq 0` always fails (1 != 0) and tests multi-arg formatting
-        let matcher = ContentMatcher::External {
-            command: vec![
-                "test".to_string(),
-                "1".to_string(),
-                "-eq".to_string(),
-                "0".to_string(),
-            ],
-        };
-        let matches = matcher.matches_with_context("content");
-        pretty_assert_eq!(matches.len(), 1);
-        // shell_words::join formats the command
-        pretty_assert_eq!(
-            matches[0].captures.get("command"),
-            Some(&"test 1 -eq 0".to_string())
-        );
-    }
-
-    #[test]
-    fn test_external_passes_content_to_stdin() {
-        // Use grep to check that specific content is passed via stdin
-        let matcher = ContentMatcher::External {
-            command: vec!["grep".to_string(), "-q".to_string(), "needle".to_string()],
-        };
-        // grep -q exits 0 if pattern found, 1 if not
-        // So if "needle" is in content, grep succeeds (no match)
-        // If "needle" is NOT in content, grep fails (match)
-        assert!(!matcher.is_match("haystack with needle inside"));
-        assert!(matcher.is_match("haystack without the search term"));
-    }
 
     #[test]
     fn test_bash_matcher_deserialize() {
@@ -1466,39 +348,5 @@ mod tests {
             serde_yaml::from_str::<PreToolUseBashMatcher>(yaml).expect("valid bash matcher yaml");
         pretty_assert_eq!(matcher.command.len(), 1);
         pretty_assert_eq!(matcher.project_state.len(), 1);
-    }
-
-    #[test]
-    fn test_project_state_git_deserialize() {
-        let yaml = r#"
-            kind: Git
-            branch:
-              - kind: Regex
-                pattern: "^main$"
-        "#;
-        let matcher = serde_yaml::from_str::<ProjectStateMatcher>(yaml)
-            .expect("valid project state matcher yaml");
-        assert!(matches!(matcher, ProjectStateMatcher::Git { .. }));
-    }
-
-    #[test]
-    fn test_project_state_git_empty_branch() {
-        let yaml = r#"
-            kind: Git
-            branch: []
-        "#;
-        let matcher = serde_yaml::from_str::<ProjectStateMatcher>(yaml)
-            .expect("valid project state matcher yaml");
-        let ProjectStateMatcher::Git { branch } = matcher;
-        assert!(branch.is_empty());
-    }
-
-    #[test]
-    fn test_project_state_invalid_kind() {
-        let yaml = r#"
-            kind: InvalidKind
-        "#;
-        let result = serde_yaml::from_str::<ProjectStateMatcher>(yaml);
-        assert!(result.is_err());
     }
 }
